authentication dan authorization

+++ latihan/app/Http/Controllers/ProdiController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Prodi;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProdiController extends Controller
 {
@@ -11,6 +12,7 @@ class ProdiController extends Controller
      * Display a listing of the resource.
      */
     public function index() {
+        Gate::authorize('viewAny', Prodi::class);
         $listprodi = Prodi::get();
         return view("prodi.index",
         ['listprodi' => $listprodi]
@@ -22,6 +24,7 @@ class ProdiController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Prodi::class);
         return view("prodi.create");
     }
 
@@ -30,6 +33,7 @@ class ProdiController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('restore', Prodi::class);
         $validateData = $request->validate([
             'nama' => 'required|min:5|max:20',
             'kode_prodi' => 'required|min:2|max:2',
@@ -55,6 +59,7 @@ class ProdiController extends Controller
      */
     public function show(string $id)
     {
+        Gate::authorize('view', Prodi::class);
         $prodi = Prodi::find($id);
         return view("prodi.detail",['prodi' => $prodi]);
 
@@ -65,6 +70,7 @@ class ProdiController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('update', Prodi::class);
         $prodi = Prodi::find($id);
         return view("prodi.edit",['prodi' => $prodi]);
     }
@@ -74,6 +80,7 @@ class ProdiController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('update', Prodi::class);
         $validateData = $request->validate([
             'nama' => 'required|min:5|max:20',
             'kode_prodi' => 'required|min:2|max:2',
@@ -94,6 +101,7 @@ class ProdiController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('delete', Prodi::class);
         $prodi = Prodi::find($id);
         $prodi->delete();
         return redirect("prodi")->with("status","data program studi berhasil dihapus");
